requirements/base.txt
# Django
Django==4.1.5
django-filter==22.1
djangorestframework==3.14.0

# Async
gunicorn==20.1.0
uvicorn[standard]==0.20.0
celery==5.2.7
redis==4.4.2
django-redis==5.2.0

# Utils
requests==2.28.1
psycopg2-binary==2.9.5
django-extensions==3.2.1
uritemplate==4.1.1
jsonschema==4.17.3
django-health-check==3.17.0
hashids==1.3.1